Rename annotation to WithJavaOptions
